edi_exchange_template: rework info provider mixins
